back end/pyproject.toml
[project]
name = "raksha-backend"
version = "0.1.0"
description = "Raksha voice assistant backend using Google ADK + FastAPI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn[standard]>=0.35.0",
  "google-adk>=1.28.0",
  "google-genai>=1.69.0",
  "pydantic-settings>=2.10.0",
  "python-dotenv>=1.1.1",
  "sqlmodel>=0.0.35",
]

[dependency-groups]
dev = [
  "pytest>=8.4.0",
  "pytest-asyncio>=1.1.0",
]

[tool.uv]
package = false

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
